% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-dataset_checking.R
\name{BridgeRDatasetChecker}
\alias{BridgeRDatasetChecker}
\title{BRIC-seq Dataset checker}
\usage{
BridgeRDatasetChecker(inputFile, group = c("Control", "Knockdown"),
  hour = c(0, 1, 2, 4, 8, 12), inforColumn = 4, percentile = c(0.99, 0.95,
  0.9, 0.8, 0.7, 0.6, 0.5, 0.4, 0.3, 0.2, 0.1, 0.05),
  outputPrefix = "BridgeR_2_raw")
}
\arguments{
\item{inputFile}{Input matrix object.}

\item{group}{The vector of group names.}

\item{hour}{The vector of time course about BRIC-seq experiment.}

\item{inforColumn}{The number of information columns.}

\item{percentile}{Percentile numbers.}

\item{outputPrefix}{The prefix for the name of the output.}
}
\description{
\code{BridgeRDatasetChecker} returns several dataset information
}

