% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{BridgeReport}
\alias{BridgeReport}
\title{Shinyapp reporting to draw RNA decay curve.}
\usage{
BridgeReport(inputFile, group = c("Control", "Knockdown"), hour = c(0, 1, 2,
  4, 8, 12), comparisonFile = c("Control", "Knockdown"),
  searchRowName = "symbol", inforColumn = 4, color = c("black", "red"),
  TimePointRemoval1 = c(1, 2), TimePointRemoval2 = c(8, 12))
}
\arguments{
\item{inputFile}{The vector of tab-delimited matrix file.}

\item{group}{The vector of group names.}

\item{hour}{The vector of time course about BRIC-seq experiment.}

\item{comparisonFile}{The vector of group names.}

\item{searchRowName}{Row name for searching.}

\item{inforColumn}{The number of information columns.}

\item{color}{color of line graph for two decay curve.}

\item{TimePointRemoval1}{The candicate_1 of time point removal.}

\item{TimePointRemoval2}{The candicate_2 of time point removal.}
}
\description{
\code{BridgeReport} returns a shinyapp to draw RNA decay curve.
}

