% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-halflife_SD.R
\name{BridgeRGrubbsTest}
\alias{BridgeRGrubbsTest}
\title{RNA half-life Grubbs test.}
\usage{
BridgeRGrubbsTest(controlFile, compFile, hour = c(0, 1, 2, 4, 8, 12),
  controlGroup = c("CTRL_PUM1", "CTRL_PUM2", "CTRL_DKD"), inforColumn = 4,
  compIndex = 2, save = T, outputPrefix = "BridgeR_8")
}
\arguments{
\item{controlFile}{The dataframe of halflife table.}

\item{compFile}{The dataframe of RPKM table.}

\item{hour}{The vector of time course about BRIC-seq experiment.}

\item{controlGroup}{The vector of group names.}

\item{inforColumn}{The number of information columns.}

\item{compIndex}{The number of information columns.}

\item{save}{Whether to save the output matrix file.}

\item{outputPrefix}{The prefix for the name of the output.}
}
\description{
\code{BridgeRGrubbsTest} returns the dataframe of
RNA half-life for each gene.
}

