% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-Normalization.R
\name{BridgeRNormalizationFactors}
\alias{BridgeRNormalizationFactors}
\title{Calculate normalization factors for BRIC-seq datasets.}
\usage{
BridgeRNormalizationFactors(inputFile, group = c("Control", "Knockdown"),
  hour = c(0, 1, 2, 4, 8, 12), inforColumn = 4, save = T, YMin = -2,
  YMax = 2, downsamplingFig = 0.2, makeFig = FALSE,
  cutoffQuantile = 0.975, figOutputPrefix = "BridgeR_3_fig_",
  factorOutputPrefix = "BridgeR_3")
}
\arguments{
\item{inputFile}{The vector of tab-delimited matrix file.}

\item{group}{The vector of group names.}

\item{hour}{The vector of time course about BRIC-seq experiment.}

\item{inforColumn}{The number of information columns.}

\item{save}{Whether to save the output matrix file.}

\item{YMin}{Y-axis min.}

\item{YMax}{Y-axis max.}

\item{downsamplingFig}{the factor for downsampling.}

\item{makeFig}{Whether to save the figure of normalization factor.}

\item{cutoffQuantile}{cutoff value of quantile.}

\item{figOutputPrefix}{The prefix for the name of figure output.}

\item{factorOutputPrefix}{The prefix for the name of factor output.}
}
\description{
\code{BridgeRNormalizationFactors} returns the dataframe of
the normalization factors for BRIC-seq datasets.
}

