% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-Normalization.R
\name{BridgeRNormalizationFactorsHK}
\alias{BridgeRNormalizationFactorsHK}
\title{Calculate normalization factors from house-keeping genes.}
\usage{
BridgeRNormalizationFactorsHK(inputFile, group = c("Control", "Knockdown"),
  hour = c(0, 1, 2, 4, 8, 12), inforColumn = 4,
  inforHKGenesRow = "symbol", HKGenes = c("GAPDH", "PGK1", "PPIA", "ENO1",
  "ATP5B", "ALDOA"), save = T, factorOutputPrefix = "BridgeR_3")
}
\arguments{
\item{inputFile}{The vector of tab-delimited matrix file.}

\item{group}{The vector of group names.}

\item{hour}{The vector of time course about BRIC-seq experiment.}

\item{inforColumn}{The number of information columns.}

\item{inforHKGenesRow}{The column number of house-keeping gene information.}

\item{HKGenes}{The vector of house-keeping genes.}

\item{save}{Whether to save the output matrix file.}

\item{factorOutputPrefix}{The prefix for the name of factor output.}
}
\description{
\code{BridgeRNormalizationFactorsHK} returns the dataframe of
the normalization factors from house-keeping genes.
}

