% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-pvalue_evaluation.R
\name{BridgeRPvalueEvaluation}
\alias{BridgeRPvalueEvaluation}
\title{Calculate Fold-change of RNA half-life and p-value.}
\usage{
BridgeRPvalueEvaluation(inputFile, group = c("Control", "Knockdown"),
  hour = c(0, 1, 2, 4, 8, 12), comparisonFile = c("Control", "Knockdown"),
  inforColumn = 4, CutoffTimePointNumber = 4, calibration = FALSE,
  save = TRUE, outputPrefix = "BridgeR_6")
}
\arguments{
\item{inputFile}{The vector of tab-delimited matrix file.}

\item{group}{The vector of group names.}

\item{hour}{The vector of time course about BRIC-seq experiment.}

\item{comparisonFile}{The vector of group names.}

\item{inforColumn}{The number of information columns.}

\item{CutoffTimePointNumber}{The number of minimum time points for calc.}

\item{calibration}{Calibration of RNA half-life.}

\item{save}{Whether to save the output matrix file.}

\item{outputPrefix}{The prefix for the name of the output.}
}
\description{
\code{BridgeRPvalueEvaluation} returns the dataframe of
Fold-change of RNA half-life and p-value.
}

