% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-RNA_halflife_comparison.R
\name{BridgeRResultChecker}
\alias{BridgeRResultChecker}
\title{BRIC-seq result checker}
\usage{
BridgeRResultChecker(inputFile, group = c("Control", "Knockdown"),
  hour = c(0, 1, 2, 4, 8, 12), inforColumn = 4, save = T,
  outputPrefix = "BridgeR_9")
}
\arguments{
\item{inputFile}{The vector of tab-delimited matrix file.}

\item{group}{The vector of group names.}

\item{hour}{The vector of time course about BRIC-seq experiment.}

\item{inforColumn}{The number of information columns.}

\item{save}{Whether to save the output fig file.}

\item{outputPrefix}{The prefix for the name of the output.}
}
\description{
\code{BridgeRResultChecker} returns  .
}

