% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{halflife_table}
\alias{halflife_table}
\title{BRIC-seq result dataset for p-value estimation using grubbs test}
\format{A data frame with 200 rows and 52 variables:
\describe{
  \item{gr_id}{Group id}
  \item{symbol}{Gene symbol}
  \item{accession_id}{Gene accession id (RefSeq)}
  \item{locus}{Genome locus}
  \item{T00_1}{RPKM value at 0h in control condition}
  \item{T01_1}{RPKM value at 1h in control condition}
  \item{T02_1}{RPKM value at 2h in control condition}
  \item{T04_1}{RPKM value at 4h in control condition}
  \item{T08_1}{RPKM value at 8h in control condition}
  \item{T12_1}{RPKM value at 12h in control condition}
  \item{Model}{RNA decay fitting model}
  \item{R2}{R2 for fitting curve}
  \item{half_life}{RNA half-life}
  \item{gr_id}{Group id}
  \item{symbol}{Gene symbol}
  \item{accession_id}{Gene accession id (RefSeq)}
  \item{locus}{Genome locus}
  \item{T00_2}{RPKM value at 0h in control condition}
  \item{T01_2}{RPKM value at 1h in control condition}
  \item{T02_2}{RPKM value at 2h in control condition}
  \item{T04_2}{RPKM value at 4h in control condition}
  \item{T08_2}{RPKM value at 8h in control condition}
  \item{T12_2}{RPKM value at 12h in control condition}
  \item{Model}{RNA decay fitting model}
  \item{R2}{R2 for fitting curve}
  \item{half_life}{RNA half-life}
  \item{gr_id}{Group id}
  \item{symbol}{Gene symbol}
  \item{accession_id}{Gene accession id (RefSeq)}
  \item{locus}{Genome locus}
  \item{T00_3}{RPKM value at 0h in control condition}
  \item{T01_3}{RPKM value at 1h in control condition}
  \item{T02_3}{RPKM value at 2h in control condition}
  \item{T04_3}{RPKM value at 4h in control condition}
  \item{T08_3}{RPKM value at 8h in control condition}
  \item{T12_3}{RPKM value at 12h in control condition}
  \item{Model}{RNA decay fitting model}
  \item{R2}{R2 for fitting curve}
  \item{half_life}{RNA half-life}
  \item{gr_id}{Group id}
  \item{symbol}{Gene symbol}
  \item{accession_id}{Gene accession id (RefSeq)}
  \item{locus}{Genome locus}
  \item{T00_4}{RPKM value at 0h in knockdown condition}
  \item{T01_4}{RPKM value at 1h in knockdown condition}
  \item{T02_4}{RPKM value at 2h in knockdown condition}
  \item{T04_4}{RPKM value at 4h in knockdown condition}
  \item{T08_4}{RPKM value at 8h in knockdown condition}
  \item{T12_4}{RPKM value at 12h in knockdown condition}
  \item{Model}{RNA decay fitting model}
  \item{R2}{R2 for fitting curve}
  \item{half_life}{RNA half-life}
}}
\usage{
halflife_table
}
\description{
A dataset containing the RPKM for each time point, information column,
RNA half-life, R2 and fitting model about 200 genes.
The variables are as follows:
}
\keyword{datasets}

