% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNA_halflife_comparison}
\alias{RNA_halflife_comparison}
\title{test BRIC-seq dataset for RNA half-life comparison}
\format{A data frame with 200 rows and 20 variables:
\describe{
  \item{gr_id}{Group id}
  \item{symbol}{Gene symbol}
  \item{accession_id}{Gene accession id (RefSeq)}
  \item{locus}{Genome locus}
  \item{CTRL_1_0h}{RPKM value at 0h in control condition}
  \item{CTRL_1_1h}{RPKM value at 1h in control condition}
  \item{CTRL_1_2h}{RPKM value at 2h in control condition}
  \item{CTRL_1_4h}{RPKM value at 4h in control condition}
  \item{CTRL_1_8h}{RPKM value at 8h in control condition}
  \item{CTRL_1_12h}{RPKM value at 12h in control condition}
  \item{gr_id}{Group id}
  \item{symbol}{Gene symbol}
  \item{accession_id}{Gene accession id (RefSeq)}
  \item{locus}{Genome locus}
  \item{KD_1_0h}{RPKM value at 0h in knockdown condition}
  \item{KD_1_1h}{RPKM value at 1h in knockdown condition}
  \item{KD_1_2h}{RPKM value at 2h in knockdown condition}
  \item{KD_1_4h}{RPKM value at 4h in knockdown condition}
  \item{KD_1_8h}{RPKM value at 8h in knockdown condition}
  \item{KD_1_12h}{RPKM value at 12h in knockdown condition}
}}
\usage{
RNA_halflife_comparison
}
\description{
A dataset containing the RPKM for each time point and information column
 about 200 genes. The variables are as follows:
}
\keyword{datasets}

