% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{BridgeRCore}
\alias{BridgeRCore}
\title{BridgeR Core function}
\usage{
BridgeRCore(inputFile, inforColumn = 4, group = c("Control", "Knockdown"),
  hour = c(0, 1, 2, 4, 8, 12), RPKMcutoff = 0.1, save = TRUE,
  outputPrefix = "BridgeR")
}
\arguments{
\item{inputFile}{The vector of tab-delimited matrix file.}

\item{inforColumn}{The number of information columns.}

\item{group}{The vector of group names.}

\item{hour}{The vector of time course about BRIC-seq experiment.}

\item{RPKMcutoff}{Cutoff value of RPKM at 0hr.}

\item{save}{Whether to save the output matrix file.}

\item{outputPrefix}{The prefix for the name of the output.}
}
\description{
\code{BridgeRCore} returns RNA half-life for each gene.
}

