% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-RNA_halflife_3models.R
\name{BridgeRHalfLifeCalc3models}
\alias{BridgeRHalfLifeCalc3models}
\title{Calculate RNA half-life for each gene.}
\usage{
BridgeRHalfLifeCalc3models(inputFile, group = c("Control", "Knockdown"),
  hour = c(0, 1, 2, 4, 8, 12), inforColumn = 4, CutoffTimePointNumber = 4,
  save = T, outputPrefix = "BridgeR_5")
}
\arguments{
\item{inputFile}{The vector of tab-delimited matrix file.}

\item{group}{The vector of group names.}

\item{hour}{The vector of time course about BRIC-seq experiment.}

\item{inforColumn}{The number of information columns.}

\item{CutoffTimePointNumber}{The number of minimum time points for calc.}

\item{save}{Whether to save the output matrix file.}

\item{outputPrefix}{The prefix for the name of the output.}
}
\description{
\code{BridgeRHalfLifeCalc3models} returns the dataframe of
RNA half-life for each gene.
}

