% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-RNA_halflife_R2_selection.R
\name{BridgeRHalfLifeCalcR2Select}
\alias{BridgeRHalfLifeCalcR2Select}
\title{Calculate RNA half-life for each gene.}
\usage{
BridgeRHalfLifeCalcR2Select(inputFile, group = c("Control", "Knockdown"),
  hour = c(0, 1, 2, 4, 8, 12), inforColumn = 4, CutoffTimePointNumber = 4,
  R2_criteria = 0.9, TimePointRemoval1 = c(1, 2), TimePointRemoval2 = c(8,
  12), ThresholdHalfLife1 = 3, ThresholdHalfLife2 = 12, save = T,
  outputPrefix = "BridgeR_5")
}
\arguments{
\item{inputFile}{The vector of tab-delimited matrix file.}

\item{group}{The vector of group names.}

\item{hour}{The vector of time course about BRIC-seq experiment.}

\item{inforColumn}{The number of information columns.}

\item{CutoffTimePointNumber}{The number of minimum time points for calc.}

\item{R2_criteria}{The cutoff of R2 for R2 selection.}

\item{TimePointRemoval1}{The candicate_1 of time point removal.}

\item{TimePointRemoval2}{The candicate_2 of time point removal.}

\item{ThresholdHalfLife1}{The cutoff of TimePointRemoval1.}

\item{ThresholdHalfLife2}{The cutoff of TimePointRemoval2.}

\item{save}{Whether to save the output matrix file.}

\item{outputPrefix}{The prefix for the name of the output.}
}
\description{
\code{BridgeRHalfLifeCalcR2Select} returns the dataframe of
RNA half-life for each gene.
}

