% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-Normalization.R
\name{BridgeRNormalization}
\alias{BridgeRNormalization}
\title{Calculate the normalized RPKM for BRIC-seq dataset.}
\usage{
BridgeRNormalization(inputFile, normFactorFile, group = c("Control",
  "Knockdown"), hour = c(0, 1, 2, 4, 8, 12), inforColumn = 4, save = T,
  outputPrefix = "BridgeR_4")
}
\arguments{
\item{inputFile}{The vector of tab-delimited matrix file.}

\item{group}{The vector of group names.}

\item{hour}{The vector of time course about BRIC-seq experiment.}

\item{inforColumn}{The number of information columns.}

\item{save}{Whether to save the output matrix file.}

\item{outputPrefix}{The prefix for the name of the output.}
}
\description{
\code{BridgeRNormalization} returns the dataframe of
the normalized RPKM values.
}

